Move LookupFactory implementations to FullAccessLookupFactories
